refactor(quiz): split create quiz code to improve readability
